refactor(web-mobile): split admin_schedule.jsx into 6 submodules (mp-d7tl)
Extracts the 1680-line admin_schedule.jsx into cohesive ES modules:
admin_schedule_utils, _pacing, _lineup, _export, _score_editor, _page.
Adds surface contract tests (Tier 1) and render smoke tests (Tier 2)
to guard against regressions during the split.

Closes mp-d7tl
